Refactor get query param filters

+++ bodzify_api/view/viewset/model/AppModelViewSet.py
@@ -1,8 +1,6 @@
 #!/usr/bin/env python
 
 import re
-from abc import abstractmethod
-import stat
 
 from django.db import IntegrityError
 from django.http import QueryDict
@@ -10,12 +8,15 @@ from rest_framework import status
 from rest_framework.exceptions import ValidationError
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
-from rest_framework.serializers import ModelSerializer
+from rest_framework.serializers import ModelSerializer, ListSerializer
 from rest_framework.exceptions import APIException
+from rest_framework import viewsets
+from typing import Union, Any, Type, Optional
 
+from bodzify_api.model.base.utils.public_standard_resource.PublicStandardResource import Fields as ModelFields
+from bodzify_api.model.base.utils.base_model.BaseModel import BaseModel
 from bodzify_api.service.Service import Service
 from bodzify_api.view import utility
-from bodzify_api.view.viewset.MultiSerializerViewSet import MultiSerializerViewSet
 
 
 class PaginatedResponseFields:
@@ -25,16 +26,21 @@ class PaginatedResponseFields:
     RESULTS = 'results'
 
 
-class AppModelViewSet(MultiSerializerViewSet):
+class AppModelViewSet(viewsets.ModelViewSet):
     permission_classes = [IsAuthenticated]
 
-    @abstractmethod
-    def _get_detailed_serializer(self, instance) -> ModelSerializer:
-        raise NotImplementedError("This method must be implemented in the subclass")
-
-    @abstractmethod
-    def _get_service(self) -> Service:
-        raise NotImplementedError("This method must be implemented in the subclass")
+    def __init__(
+            self,
+            service: Optional[Service],
+            model_class: type[BaseModel],
+            filter_class: Optional[Any] = None,
+            detailed_serializer_class: Optional[Type[ModelSerializer]] = None,
+            **kwargs):
+        super().__init__(**kwargs)
+        self.service = service
+        self.model_class = model_class
+        self.filter_class = filter_class
+        self.detailed_serializer_class = detailed_serializer_class
 
     @staticmethod
     def camel_to_snake(name):
@@ -65,15 +71,27 @@ class AppModelViewSet(MultiSerializerViewSet):
             snake_case_dict[snake_case_key] = value
         return snake_case_dict
 
-    def __init__(self, service: Service, **kwargs):
-        super().__init__(**kwargs)
-        self.service = service
+    def _get_detailed_serializer_class(self) -> Type[ModelSerializer]:
+        if not self.detailed_serializer_class:
+            raise NotImplementedError(
+                'You must define detailed_serializer_class in your viewset initialization'
+            )
+        return self.detailed_serializer_class
+
+    def _get_detailed_serializer_instance(self, instance) -> Union[ListSerializer, ModelSerializer, Any]:
+        serializer_class = self._get_detailed_serializer_class()
+        if serializer_class is not None:
+            return serializer_class(instance=instance)
+        raise ValidationError("Serializer class not defined")
 
     def _create(self, request, *args, **kwargs):
+        if not self.service:
+            raise NotImplementedError("Service not defined in viewset")
+
         request_data_snake_case = self.get_dict_with_snake_case_keys_from_form_data(request.data)
         try:
             instance = self.service.create(post_data=request_data_snake_case, request=request)
-            response_serializer = self._get_detailed_serializer(instance=instance)
+            response_serializer = self._get_detailed_serializer_instance(instance=instance)
             headers = self.get_success_headers(response_serializer.data)
             return Response(data=response_serializer.data, status=status.HTTP_201_CREATED, headers=headers)
 
@@ -81,27 +99,52 @@ class AppModelViewSet(MultiSerializerViewSet):
             return utility.get_response_when_bad_request(exception=exception)
 
     def _update(self, request, *args, **kwargs):
+        if not self.service:
+            raise NotImplementedError("Service not defined in viewset")
+
         request_data_snake_case = self.get_dict_with_snake_case_keys_from_form_data(request.data)
         updatedinstance = self.service.update(put_data=request_data_snake_case,
                                               oldinstance=self.get_object(),
                                               request=request)
-        response_serializer_data = self._get_detailed_serializer(updatedinstance).data
-        headers = self.get_success_headers(response_serializer_data)
-        return Response(data=response_serializer_data, status=status.HTTP_200_OK, headers=headers)
+        response_serializer = self._get_detailed_serializer_instance(updatedinstance)
+        headers = self.get_success_headers(response_serializer.data)
+        return Response(data=response_serializer.data, status=status.HTTP_200_OK, headers=headers)
 
-    def _list(self, request, *args, **kwargs):
-        queryset = self.get_queryset()
+    def get_queryset(self):
+        try:
+            snake_case_params = self.get_dict_in_snake_case_keys_from_dict_in_camel_case_keys(self.request.GET)
+            queryset = self.model_class.objects.filter(user=self.request.user)
 
-        page = self.paginate_queryset(queryset)
-        if page:
-            data = self.get_serializer(page, many=True).data
+            if self.filter_class:
+                queryset = self.filter_class(snake_case_params, queryset=queryset).qs
 
-        if not queryset.exists():
-            data = []
+            ordering_fields = self.model_class.objects.get_default_ordering()
+            return queryset.order_by(*ordering_fields)
+        except ValidationError as e:
+            raise ValidationError(e.detail)
 
-        return self.get_paginated_response(data)
+    def list(self, request, *args, **kwargs):
+        try:
+            queryset = self.get_queryset()
+
+            page = self.paginate_queryset(queryset)
+            if page:
+                data = self._get_detailed_serializer_class()(page, many=True).data
+
+            if not queryset.exists():
+                data = []
+
+            return self.get_paginated_response(data)
+        except ValidationError as e:
+            return Response(
+                {"detail": str(e.detail[0])},
+                status=status.HTTP_400_BAD_REQUEST
+            )
 
     def _destroy(self, request, *args, **kwargs):
+        if not self.service:
+            raise NotImplementedError("Service not defined in viewset")
+
         self.service.delete(user=request.user, instance=self.get_object())
         return Response(status=status.HTTP_204_NO_CONTENT)
 
@@ -126,5 +169,5 @@ class AppModelViewSet(MultiSerializerViewSet):
 
     def retrieve(self, request, *args, **kwargs):
         instance = self.get_object()
-        serializer = self.get_serializer(instance)
+        serializer = self._get_detailed_serializer_instance(instance)
         return Response(serializer.data)
